pyproject.toml
[tool.poetry]
name = "ditto"
version = "0.2.3"
description = "Distribution Feeder Conversion Tool"
license = "BSD license"
keywords = ["ditto"]
classifiers = ["Development Status :: 2 - Pre-Alpha", "Intended Audience :: Developers", "License :: OSI Approved :: BSD License", "Natural Language :: English", "Programming Language :: Python :: 2.7", "Programming Language :: Python :: 3.5", "Programming Language :: Python :: 3.6"]
homepage = "https://github.com/NREL/ditto"
authors = ["Tarek Elgindy <tarek.elgindy@nrel.gov>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
click = "^8.1.3"
future = "^0.18.3"
json-tricks = "^3.16.1"
networkx = "^3.1"
pandas = "^1.5.3"
scipy = "^1.10.1"
six = "^1.16.0"
traitlets = ">=4.1"
tqdm = "^4.65.0"

ghp-import = { optional = true, version = "*" }
"backports.tempfile" = { optional = true, version = "*" }
croniter = { optional = true, version = "*" }
lxml = { optional = true, version = "*" }
nbsphinx = { optional = true, version = "*" }
numpy = { version = ">=1.13.0", optional = true }
"opendssdirect.py" = { version = ">=0.3.3", optional = true }
pandas-access = { optional = true, version = "*" }
pytest = { optional = true, version = "*" }
pytest-cov = { optional = true, version = "*" }
sphinx-rtd-theme = { optional = true, version = "*" }
sphinxcontrib-napoleon = { optional = true, version = "*" }
xlrd = { optional = true, version = "*" }
xlsxwriter = { optional = true, version = "*" }

[tool.poetry.dev-dependencies]
black = "*"
pre-commit = "*"
pypandoc = "*"

[tool.poetry.extras]
test = ["backports-tempfile", "ghp-import", "nbsphinx", "pytest", "pytest-cov", "sphinx-rtd-theme", "sphinxcontrib-napoleon"]
dew = ["numpy", "xlrd"]
synergi = ["numpy", "pandas"]
cyme = ["numpy"]
gridlabd = ["croniter", "numpy"]
ephasor = ["numpy", "pandas"]
all = ["croniter", "lxml", "numpy", "opendssdirect-py", "pandas", "pandas-access", "scipy", "xlrd", "xlsxwriter"]
extras = ["croniter", "lxml", "numpy", "opendssdirect-py", "pandas", "pandas-access", "scipy", "xlrd", "xlsxwriter"]
opendss = ["OpenDSSDirect.py", "pandas", "numpy"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-ordering = "^0.6"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
ditto = "ditto.cli:main"
ditto-cli = "ditto.cli:cli"

[tool.pytest.ini_options]
# Ignore library warnings - they just pollute the test output
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning", "ignore::FutureWarning"]

# default options added when running `poetry run pytest`:
#addopts = '-s -m "not integration" --cov-report term-missing --cov-report html --cov=src --durations=0 --durations-min=5.0 --tb=native'

# options to let debuggers work with pytest (by disabling pytest-cov coverage, see https://youtrack.jetbrains.com/issue/PY-20186)
addopts = '-p no:faulthandler --durations=0 --durations-min=5.0 --tb=native'


